Factor out some things
