+++ recipes/default.rb
@@ -24,70 +24,34 @@
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #
 
-# so it turns out I didn't remember how to recursively walk a nested hash like this :/
-def walk hash, stack = [], output = {}
-  hash.each do |key, val|
-    stack << key
-    if val.is_a?(Hash)
-      walk val, stack, output
-    else
-      output[stack.join(node["envbuilder"]["joiner"]).upcase] = val
-      stack.pop
-    end
-  end
-  stack.pop
-
-  output
+class Chef::Recipe
+  include ODI::Envbuilder::Helpers
 end
 
-def dump_hash hash
-  s = ''
-  hash.each do |key, val|
-    s << "%s: %s\n" % [
-        key,
-        val
-    ]
-  end
-
-  s
-end
-
-def use_encrypted_data_bag?
-  !!node["envbuilder"]["use_encrypted_data_bag"]
-end
-
-def bag_item(bag_name, item_name)
-  if use_encrypted_data_bag?
-    Chef::EncryptedDataBagItem.load(bag_name, item_name)
-  else
-    data_bag_item(bag_name, item_name)
-  end
-end
-
-default_item = bag_item(
+default_data = walk bag_item(
   node["envbuilder"]["data_bag"],
   node["envbuilder"]["base_dbi"]
-)
+)['content']
 
-environment_item = bag_item(
+environment_data = walk bag_item(
   node["envbuilder"]["data_bag"],
   node["ENV"] || node.chef_environment
-)
+)['content']
 
 master_list = walk bag_item(
   node["envbuilder"]["data_bag"],
   'master_list'
 )['content']
 
-merged_item = (walk default_item["content"]).update (walk environment_item["content"])
+merged_data = (default_data).update (environment_data)
 
-merged_item.each do |k, v|
+merged_data.each do |k, v|
   if v == 'DEFAULT'
-    merged_item[k] = master_list[k]
+    merged_data[k] = master_list[k]
   end
 end
 
-content = dump_hash merged_item
+content = dump_hash merged_data
 
 group node["envbuilder"]["group"] do
   action :create
